safari-600.7-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.8_Literals/7.8.3_Numeric_Literals/S7.8.3_A6.1_T1-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: No hexadecimal digits after '0x'
S7.8.3_A6.1_T1

PASS Expected parsing failure

TEST COMPLETE
